(split)bug: none (checkpoint)
Copied from Perforce
 Change: 27161

+++ src/java/com/zimbra/cs/zclient/ZFolder.java
@@ -89,7 +89,7 @@ public interface ZFolder {
      * 
      * @return color
      */
-    public byte getColor();
+    public int getColor();
 
     /**
      * remote URL (RSS, iCal, etc) this folder syncs to
